add migration for unique author and name, and textlimit, and refactor repositories
