+++ Chirp.Razor/Infrastructure/MessageRepository.cs
@@ -1,18 +1,7 @@
 using Chirp.Razor.Models;
 using Microsoft.EntityFrameworkCore;
 
-namespace Chirp.Razor.Data;
-
-
-// Keep a simple, slide-friendly shape.
-// Methods exist but we won’t implement them yet.
-public interface IMessageRepository
-{
-    Task<List<MessageDTO>> ReadMessagesAsync(string? author = null, int page = 0, int pageSize = 32);
-    Task<int> CreateMessageAsync(MessageDTO message);
-    Task UpdateMessageAsync(MessageDTO message);
-    Task DeleteMessageAsync(int id);
-}
+namespace Chirp.Razor.Infrastructure;
 
 public class MessageRepository : IMessageRepository
 {
